requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
django-tempus-dominus==5.1.2.16
django-widget-tweaks==1.4.9
gunicorn==20.1.0
psycopg2==2.9.2
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
